requirements.txt
# Minecraft Server Wrapper Dependencies
# ======================================
# Install with: pip install -r requirements.txt

# PyYAML for configuration file parsing
PyYAML>=6.0
